requirements.txt
Pillow
QuantLib
arch
beautifulsoup4
boruta
browser_cookie3
hmmlearn
investpy
ipython
kiteconnect
matplotlib
minisom
mplfinance
numpy
pandas
pandas-ta
pandas_market_calendars
pendulum
plotly
python-dateutil
python-dotenv
pytz
requests
schedule
scikit-learn
scipy
seaborn
selenium
shap
statsmodels
streamlit
tenacity
tqdm
vectorbt
yfinance
